# projects/points-web-app/docs/v0.2/points-markets.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Freeism Points–Markets API",
    "version": "0.2.0",
    "description": "Points-owned canonical contract for the twelve APIs consumed by Markets. Better Auth discovery, authorization, token, introspection and revocation endpoints remain standard provider endpoints and are intentionally not duplicated as application operations."
  },
  "servers": [
    {
      "url": "https://points.freeism.app"
    }
  ],
...
  "paths": {
...
    "/api/v1/point-package-auction-eligibility-checks": {
      "post": {
        "operationId": "checkPointPackageAuctionEligibility",
        "summary": "Check Point Package auction eligibility atomically",
        "description": "Checks all items in one Points D1 transaction. One ineligible item produces zero receipts and a safe sorted error list.",
        "security": [
          {
            "m2mOAuth": ["points.packages.auction-eligibility"]
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Opaque idempotency key. The same key and canonical payload replay the initial HTTP status and domain result.",
            "schema": {
              "$ref": "#/components/schemas/OpaqueId"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/OpaqueId"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuctionEligibilityRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "headers": {
              "Cache-Control": {
                "description": "Exact cache policy: private, no-store",
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "private, no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionEligibilityResponse"
                }
              }
            }
          },
          "409": {
            "description": "Auction ineligibility or idempotency-key reuse conflict. No eligibility receipt is issued.",
            "headers": {
              "Cache-Control": {
                "description": "Exact cache policy: private, no-store",
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "private, no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/AuctionEligibilityProblem"
                    },
                    {
                      "$ref": "#/components/schemas/IdempotencyKeyReusedProblem"
                    }
                  ]
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "default": {
            "description": "RFC 9457 Problem Details",
            "headers": {
              "Cache-Control": {
                "description": "Exact cache policy: private, no-store",
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "private, no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-freeism-request-body-max-bytes": 1048576,
        "x-freeism-idempotency": {
          "replayHttpStatus": "preserve-initial",
          "replayDomainResult": true,
          "requestIdMayBeRegenerated": true,
          "sameKeyDifferentPayloadStatus": 409,
          "sameKeyDifferentPayloadCode": "IDEMPOTENCY_KEY_REUSED"
        },
        "x-freeism-cache-control": "private, no-store",
        "x-freeism-receipt": {
          "leaseSeconds": 30,
          "validUntilExclusive": true,
          "boundTo": [
            "clientId",
            "auctionCommandId",
            "auctionCommandHash",
            "allItems.auctionItemId",
            "allItems.pointPackageId",
            "allItems.pointPackageRevisionId",
            "allItems.contentHash"
          ],
          "allOrNothing": true,
          "failureReceiptCount": 0,
          "idempotentReplayExtendsLease": false
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "userOAuth": {
        "type": "oauth2",
        "description": "JWT Authorization Code + PKCE access tokens for ordinary user operations. Refresh tokens may renew only this user scope set.",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://points.freeism.app/api/auth/oauth2/authorize",
            "tokenUrl": "https://points.freeism.app/api/auth/oauth2/token",
            "refreshUrl": "https://points.freeism.app/api/auth/oauth2/token",
            "scopes": {
              "openid": "OpenID Connect identity",
              "profile": "Basic profile",
              "offline_access": "Refresh token access",
              "points.connection.read": "Read the active Points connection",
              "points.balance.read": "Check point balances",
              "points.reservations.create": "Create point reservations"
            },
            "x-freeism-pkce": {
              "required": true,
              "method": "S256"
            }
          }
        },
        "x-freeism-access-token-format": "jwt",
        "x-freeism-principal-class": "user"
      },
      "connectionUnlinkOAuth": {
        "type": "oauth2",
        "description": "One-time Authorization Code + PKCE token after Google fresh authentication for normal unlink only.",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://points.freeism.app/api/auth/oauth2/authorize",
            "tokenUrl": "https://points.freeism.app/api/auth/oauth2/token",
            "scopes": {
              "points.connection.unlink": "Deactivate exactly one active Points connection"
            },
            "x-freeism-pkce": {
              "required": true,
              "method": "S256"
            }
          }
...
    "responses": {
      "RateLimited": {
        "description": "RFC 9457 rate limit Problem Details",
        "headers": {
          "Cache-Control": {
            "description": "Exact cache policy: private, no-store",
            "required": true,
            "schema": {
              "type": "string",
              "const": "private, no-store"
            }
          },
          "Retry-After": {
            "description": "Non-empty HTTP Retry-After value",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/RateLimitedProblem"
            }
          }
        }
      }
    },
    "schemas": {
      "OpaqueId": {
        "type": "string",
        "minLength": 1,
        "maxLength": 255
      },
...
      "Sha256Hash": {
        "type": "string",
        "pattern": "^sha256:[0-9a-f]{64}$"
      },
      "UtcInstant": {
        "type": "string",
        "format": "date-time"
      },
...
      "PositiveSafeInteger": {
        "type": "integer",
        "minimum": 1,
        "maximum": 9007199254740991
      },
      "NonNegativeSafeInteger": {
        "type": "integer",
        "minimum": 0,
        "maximum": 9007199254740991
      },
...
      "RequestMeta": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/OpaqueId"
          }
        },
        "required": ["requestId"]
      },
      "ValidationError": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "row": {
            "$ref": "#/components/schemas/NonNegativeSafeInteger"
          },
          "field": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "code": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]*$"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          }
        },
        "required": ["code"]
      },
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "instance": {
            "type": "string",
            "format": "uri-reference"
          },
          "code": {
            "type": "string",
            "enum": [
              "MALFORMED_REQUEST",
              "AUTHENTICATION_REQUIRED",
              "INVALID_ACCESS_TOKEN",
              "INSUFFICIENT_SCOPE",
              "RESOURCE_NOT_FOUND",
              "CONTENT_TYPE_UNSUPPORTED",
              "REQUEST_BODY_TOO_LARGE",
              "VALIDATION_FAILED",
              "IDEMPOTENCY_KEY_REQUIRED",
              "IDEMPOTENCY_KEY_REUSED",
              "RATE_LIMITED",
              "INTERNAL_ERROR",
              "DEPENDENCY_UNAVAILABLE",
              "POINT_PACKAGE_AUCTION_INELIGIBLE",
              "LINK_ATTEMPT_ALREADY_FINALIZED",
              "ACTIVE_RESERVATION_EXISTS",
              "INSUFFICIENT_BALANCE",
              "POINT_RESERVATION_NOT_ACTIVE",
              "POINT_RESERVATION_EXPIRED",
              "RESERVATION_VECTOR_HASH_MISMATCH",
              "SETTLEMENT_PLAN_HASH_MISMATCH"
            ]
          },
          "requestId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "minItems": 1
          }
        },
        "required": ["type", "title", "status", "code", "requestId"]
      },
      "RateLimitedProblem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "type": "integer",
            "const": 429
          },
          "detail": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "instance": {
            "type": "string",
            "format": "uri-reference"
          },
          "code": {
            "type": "string",
            "const": "RATE_LIMITED"
          },
          "requestId": {
            "$ref": "#/components/schemas/OpaqueId"
          }
        },
        "required": ["type", "title", "status", "code", "requestId"]
      },
...
      "AuctionEligibilityRequestItem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "auctionItemId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "pointPackageId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "pointPackageRevisionId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "contentHash": {
            "$ref": "#/components/schemas/Sha256Hash"
          }
        },
        "required": ["auctionItemId", "pointPackageId", "pointPackageRevisionId", "contentHash"]
      },
      "AuctionEligibilityRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "auctionCommandId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "auctionCommandHash": {
            "$ref": "#/components/schemas/Sha256Hash"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuctionEligibilityRequestItem"
            },
            "minItems": 1,
            "maxItems": 1000,
            "x-freeism-unique-by": "auctionItemId"
          }
        },
        "required": ["auctionCommandId", "auctionCommandHash", "items"]
      },
      "AuctionEligibilityReceiptItem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "auctionItemId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "pointPackageId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "pointPackageRevisionId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "contentHash": {
            "$ref": "#/components/schemas/Sha256Hash"
          },
          "packageEligibilityVersion": {
            "$ref": "#/components/schemas/PositiveSafeInteger"
          }
        },
        "required": [
          "auctionItemId",
          "pointPackageId",
          "pointPackageRevisionId",
          "contentHash",
          "packageEligibilityVersion"
        ]
      },
      "AuctionEligibilityData": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "pointPackageAuctionEligibilityReceiptId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "auctionCommandId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "auctionCommandHash": {
            "$ref": "#/components/schemas/Sha256Hash"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuctionEligibilityReceiptItem"
            },
            "minItems": 1,
            "maxItems": 1000,
            "x-freeism-sort": ["auctionItemId:asc"]
          },
          "checkedAt": {
            "$ref": "#/components/schemas/UtcInstant"
          },
          "validUntil": {
            "$ref": "#/components/schemas/UtcInstant"
          }
        },
        "required": [
          "pointPackageAuctionEligibilityReceiptId",
          "auctionCommandId",
          "auctionCommandHash",
          "items",
          "checkedAt",
          "validUntil"
        ]
      },
      "AuctionEligibilityResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AuctionEligibilityData"
          },
          "meta": {
            "$ref": "#/components/schemas/RequestMeta"
          }
        },
        "required": ["data", "meta"]
      },
      "AuctionEligibilityItemError": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "auctionItemId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "code": {
            "type": "string",
            "enum": [
              "POINT_PACKAGE_NOT_FOUND",
              "POINT_PACKAGE_REVISION_NOT_FOUND",
              "POINT_PACKAGE_REVISION_MISMATCH",
              "POINT_PACKAGE_REVISION_INACTIVE",
              "POINT_PACKAGE_INACTIVE",
              "CONTENT_HASH_MISMATCH"
            ]
          }
        },
        "required": ["auctionItemId", "code"]
      },
      "AuctionEligibilityProblem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "type": "integer",
            "const": 409
          },
          "code": {
            "type": "string",
            "const": "POINT_PACKAGE_AUCTION_INELIGIBLE"
          },
          "requestId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "detail": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "instance": {
            "type": "string",
            "format": "uri-reference"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuctionEligibilityItemError"
            },
            "minItems": 1,
            "x-freeism-sort": ["auctionItemId:asc"]
          }
        },
        "required": ["type", "title", "status", "code", "requestId", "errors"]
      },
      "IdempotencyKeyReusedProblem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "type": "integer",
            "const": 409
          },
          "code": {
            "type": "string",
            "const": "IDEMPOTENCY_KEY_REUSED"
          },
          "requestId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "detail": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "instance": {
            "type": "string",
            "format": "uri-reference"
          }
        },
        "required": ["type", "title", "status", "code", "requestId"]
      },
...
    }
  }
}